#8 공통 컴포넌트 개발 - Template - Header, Footer, Template 컴포넌트 생성
