--- frontend/components/Header/Header.scss
@@ -1,3 +1,0 @@
-.header {
-  background-color: $oc-gray-3;
-}
