+++ frontend/components/common/Footer/Footer.scss
@@ -1,0 +1,3 @@
+.footer {
+  border: 1px solid $oc-pink-6;
+}
