+++ frontend/components/common/Header/Header.scss
@@ -1,0 +1,7 @@
+.header {
+  position: fixed;
+  top: 0;
+  width: 100%;
+  height: $header-height;
+  border: 1px solid $oc-yellow-6;
+}
