+++ frontend/components/common/PageTemplate/PageTemplate.scss
@@ -1,0 +1,4 @@
+.page-template {
+  margin-top: $header-height;
+  border: 1px solid $oc-lime-6;
+}
