+++ frontend/pages/index.tsx
@@ -1,9 +1,9 @@
-import Layout from "../components/Layout/Layout";
+import PageTemplate from "../components/common/PageTemplate";
 
 const Index = props => (
-  <Layout>
-    index 페이지
-  </Layout>
+  <PageTemplate>
+    <div>index 페이지</div>
+  </PageTemplate>
 );
 
 export default Index;
