+++ frontend/styles/utils/_variables.scss
@@ -1,1 +1,2 @@
 $breakpoints: (small: 320px, medium: 768px, large: 1024px, wide: 1400px);
+$header-height: 2.7rem;
